user-insights/requirements.txt
Flask==2.1.3
requests==2.26.0
PyYAML==6.0
sqlite3
boto3==1.26.0  
mysql-connector-python==8.0.33